refactor: 수업 진행을 위한 클래스들 패키지로 이동
